Golang version (#134)
* Bump golang version to 1.20

* Remove args from testcache

Fix flakey tests

* Test changes are coming during test rewrite

Temporarily missing coverage isn't good, but better than having unpredictable tests.

* Ignore issue with defer

This looks like it could be an issue in limited situations when writing to disk. However I didn't see an example of harm to network connections, and believe this is something a user could replay if there was an issue.